+++ src/osc/commands/DefLoad.cpp
@@ -31,7 +31,7 @@ void DefLoad::processMessage(int argc, lo_arg** argv, const char* types, lo_addr
         }
     }
     std::shared_ptr<Address> origin(new Address(address));
-    m_dispatcher->async()->scinthDefLoadFile(path, [this, origin, onCompletion](int) {
+    m_dispatcher->async()->scinthDefLoadFile(path, [this, origin, onCompletion](size_t) {
         if (onCompletion) {
             m_dispatcher->processMessageFrom(origin->get(), onCompletion);
         }
